Add some simple div-free tests for extruded intervals
